Queue command

+++ src/Commands/Voice/Leave.js
@@ -1,7 +1,5 @@
 const Command = require("../../Structures/Command.js")
 
-const deleteMessage = m => m.delete(60000)
-
 module.exports = class Play extends Command {
     constructor(client, id) {
         super(client, id)
@@ -12,7 +10,6 @@ module.exports = class Play extends Command {
         }
         this.aliases = ["disconnect"]
         this.channels = ["text"]
-        this.requiredPermissions = "MANAGE_MESSAGES"
 
         this.types = {
             "force": {}
@@ -20,15 +17,16 @@ module.exports = class Play extends Command {
 
     }
 
-    async run(msg, params, { voice }) {
-        msg.delete()
+    async run(msg, params, { voice, deleteMessage }) {
+        if(msg.channel.permissions.has("MANAGE_MESSAGES"))
+            msg.delete()
         if(!voice.connection) 
             return msg.channel.send("I'm not in voice channel!")
 
         if(params.force && !(msg.author.id === msg.guild.ownerID || await msg.author.hasPermission(this.client.constants.PERMISSIONS.FORCE, msg.guild))) 
-            return msg.channel.send(`Sorry, you cannot do that.`).then(deleteMessage)
+            return msg.reply(`Sorry, you cannot do that.`).then(deleteMessage)
         else if(!params.force && voice.playing)
-            return msg.channel.send(`I can't leave while I'm playing a song.`).then(deleteMessage)
+            return msg.reply(`I can't leave while I'm playing a song.`).then(deleteMessage)
 
         voice.leave()
     }
